Move document_symbol to new requests module
