+++ vhdl_ls/src/requests/document_symbol.rs
